requirements.txt
PyYAML
gensim
gsitk
nltk
numpy
pandas
scipy
sklearn
spacy
tqdm
